Snapshot: Refactored to make toolmaking easier
